requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
requests==2.31.0
feedparser==6.0.10
python-multipart==0.0.6
python-dotenv==1.0.0 